backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
PyJWT[crypto]>=2.9.0
cryptography>=44.0.0
boto3>=1.35.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
slowapi>=0.1.9
certifi
python-multipart
